Create top-level `plan`; rename all CNF-compiling functions from `from_cnf` to `compile_cnf`; move compilation up to trait level
