+++ examples/semantic_hash_experiment.rs
@@ -119,12 +119,12 @@ impl Display for BenchStats {
 fn run_compr_sem(cnf: &Cnf, vtree: &VTree) -> (BenchStats, BenchStats) {
     let start = Instant::now();
     let compr_builder = CompressionSddBuilder::new(vtree.clone());
-    let compr_cnf = compr_builder.from_cnf(cnf);
+    let compr_cnf = compr_builder.compile_cnf(cnf);
     let compr = BenchStats::from_run("c".to_owned(), start.elapsed(), &compr_cnf, &compr_builder);
 
     let start = Instant::now();
     let sem_builder = SemanticSddBuilder::<479001599>::new(vtree.clone());
-    let sem_cnf = sem_builder.from_cnf(cnf);
+    let sem_cnf = sem_builder.compile_cnf(cnf);
     let sem = BenchStats::from_run("s".to_owned(), start.elapsed(), &sem_cnf, &sem_builder);
 
     (compr, sem)
@@ -188,7 +188,7 @@ fn run_canonicalizer_experiment(c: Cnf, vtree: VTree, verbose: bool) {
     let start = Instant::now();
 
     let compr_builder = CompressionSddBuilder::new(vtree.clone());
-    let compr_cnf = compr_builder.from_cnf(&c);
+    let compr_cnf = compr_builder.compile_cnf(&c);
 
     println!(" ");
 
@@ -201,7 +201,7 @@ fn run_canonicalizer_experiment(c: Cnf, vtree: VTree, verbose: bool) {
 
     // other primes: 100000049, 18_446_744_073_709_551_591
     let sem_builder = SemanticSddBuilder::<18_446_744_073_709_551_591>::new(vtree);
-    let sem_cnf = sem_builder.from_cnf(&c);
+    let sem_cnf = sem_builder.compile_cnf(&c);
 
     println!(
         "{}\n",
